Unit tests to verify relative_time were not calling the parse_time function in query.cc as expected. Changing the way we post query made the the function to be called properly. Also the UT has been added to check OBJECT_VALUE_TABLE. Code changes: Query class has been moved from perftest.py to analytics_fixture.py as it is used elsewhere and util would be a proper place. Instead of query query_json objects are called so that passed parameters are not interpreted by the util scripts before calling the actual functions. Test: Coverage report shows the parse_time function being referred in query.cc
